# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Validation Center API
  version: 0.1.0
  description: >
    Phase 0 + Phase 1 contract for the AI-assisted Data Comparison /
    Reconciliation Platform (deterministic core). AI suggestion (P4) and PDF
    extraction (P3) endpoints are reserved and not included here.
    Principle: AI suggests, humans approve, the rule engine judges.
servers:
  - url: /api/v1

# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
...
paths:
...
  /datasets:
    post:
      tags: [Datasets]
      summary: Upload a Source/Destination dataset (CSV/Excel). Async profiling.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/DatasetRole' }
                name: { type: string }
                file: { type: string, format: binary }
      responses:
        '202':
          description: Accepted (profiling queued)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
...
  /recon/runs:
    post:
      tags: [Reconciliation]
      summary: Trigger a reconciliation run (async batch). Requires OPERATOR role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '202':
          description: Accepted (queued)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '422': { description: Inputs not in runnable state }
    get:
      tags: [Reconciliation]
      summary: List runs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Run' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Dataset:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/DatasetRole' }
        name: { type: string }
        origin: { type: string, enum: [CSV, EXCEL] }
        rowCount: { type: integer, nullable: true }
        status: { type: string, enum: [UPLOADED, PROFILED, READY, FAILED] }
...
    RunCreate:
      type: object
      required: [templateVersionId, mappingId, sourceDatasetId, destDatasetId]
      properties:
        templateVersionId: { type: string, format: uuid }
        mappingId: { type: string, format: uuid }
        sourceDatasetId: { type: string, format: uuid }
        destDatasetId: { type: string, format: uuid }
    Run:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED] }
        ruleEngineVersion: { type: string }
        queuedAt: { type: string, format: date-time }
...
    DatasetRole:
      type: string
      enum: [SOURCE, DESTINATION]
...